Integrating the Async example into AdePT
