+++ examples/AsyncExample/CMakeLists.txt
@@ -1,6 +1,8 @@
 # SPDX-FileCopyrightText: 2023 CERN
 # SPDX-License-Identifier: Apache-2.0
 
+SET(CMAKE_EXE_LINKER_FLAGS "-Wl,--disable-new-dtags")
+
 if(HepMC3_FOUND)
   message(STATUS "HepMC3 found ${HEPMC3_INCLUDE_DIR}")
 else()
@@ -71,36 +73,15 @@ endif()
 # Include the C++ side of the AdePT-G4 integration from core AdePT:
 list(TRANSFORM ADEPT_G4_INTEGRATION_SRCS PREPEND ../../)
 
-add_library(AdePT_G4_integration_async SHARED
-  ${ADEPT_G4_INTEGRATION_SRCS}
-  AdeptIntegration.cpp
-  AdeptIntegration.cu
-  BasicScoring.cu
-  ResourceManagement.cu)
-target_link_libraries(AdePT_G4_integration_async
-PUBLIC
-  CopCore
-  VecGeom::vgdml
-  VecGeom::vecgeomcuda_static
-  G4HepEm::g4HepEm
-  )
-target_include_directories(AdePT_G4_integration_async
-  PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/>
-)
-set_target_properties(AdePT_G4_integration_async PROPERTIES
-  CUDA_SEPARABLE_COMPILATION ON
-  CUDA_RESOLVE_DEVICE_SYMBOLS ON)
-
-option(AsyncTransport "Link to AdePT_G4_integration_async" ON)
-
 add_executable(AsyncExample
     example.cpp
     ${sources_g4})
 target_include_directories(AsyncExample PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/include
-  ${CMAKE_CURRENT_SOURCE_DIR})
+  ${CMAKE_CURRENT_SOURCE_DIR}
+  #TEMPORARY WORKAROUND
+  ${PROJECT_SOURCE_DIR}/include
+)
 
 target_link_libraries(AsyncExample
   PRIVATE
@@ -111,11 +92,7 @@ target_link_libraries(AsyncExample
     ROOT::RIO
 )
 
-if(AsyncTransport)
-  target_link_libraries(AsyncExample PRIVATE AdePT_G4_integration_async)
-else()
-  target_link_libraries(AsyncExample PRIVATE AdePT_G4_integration)
-endif()
+target_link_libraries(AsyncExample PRIVATE AdePT_G4_integration)
 
 target_compile_options(AsyncExample PRIVATE -Wall -Wextra)
