+++ examples/CMakeLists.txt
@@ -9,4 +9,4 @@ configure_file(data/ppttbar.hepmc3 ${PROJECT_BINARY_DIR}/ppttbar.hepmc3)
 # - Subprojects
 add_subdirectory(Example1)
 add_subdirectory(IntegrationBenchmark)
-# add_subdirectory(AsyncExample)
+add_subdirectory(AsyncExample)
