+++ include/AdePT/base/SlotManager.cuh
