+++ include/AdePT/core/AsyncAdePTTransport.hh
@@ -6,15 +6,18 @@
 /// - filling the buffer with tracks to be transported on the GPU
 /// - Calling the Shower method transporting a buffer on the GPU
 
-#ifndef ADEPT_INTEGRATION_H
-#define ADEPT_INTEGRATION_H
+#ifndef ASYNC_ADEPT_TRANSPORT_HH
+#define ASYNC_ADEPT_TRANSPORT_HH
 
 #define ADEPT_SAVE_IDs
-#include "TrackTransfer.h"
-#include "BasicScoring.h"
+// #include "TrackTransfer.h"
+// #include "BasicScoring.h"
 
 #include <AdePT/core/AdePTTransportInterface.hh>
 #include <AdePT/core/CommonStruct.h>
+#include <AdePT/core/AdePTConfiguration.hh>
+#include <AdePT/core/AsyncAdePTTransportStruct.hh>
+#include <AdePT/core/PerEventScoringStruct.cuh>
 
 #include <VecGeom/base/Config.h>
 #include <VecGeom/management/CudaManager.h> // forward declares vecgeom::cxx::VPlacedVolume
@@ -25,45 +28,38 @@
 #include <thread>
 #include <unordered_map>
 
+// struct AdePTScoring;
+
 class G4Region;
 class G4VPhysicalVolume;
 struct G4HepEmState;
-class AdePTGeant4Integration;
+// class AdePTGeant4Integration;
 namespace AsyncAdePT {
 struct TrackBuffer;
 struct GPUstate;
-struct HitProcessingContext;
+// enum class EventState : unsigned char;
+// struct HitProcessingContext;
+
 void InitVolAuxArray(adeptint::VolAuxArray &array);
 
-class AdeptIntegration : public AdePTTransportInterface {
+template <typename IntegrationLayer>
+class AsyncAdePTTransport : public AdePTTransportInterface {
 public:
-  static constexpr int kMaxThreads = 256;
   static inline uint64_t fAdePTSeed = 1234567;
 
 private:
-  enum class EventState : unsigned char {
-    NewTracksFromG4,
-    G4RequestsFlush,
-    Inject,
-    InjectionCompleted,
-    Transporting,
-    WaitingForTransportToFinish,
-    RequestHitFlush,
-    FlushingHits,
-    HitsFlushed,
-    FlushingTracks,
-    DeviceFlushed,
-    LeakedTracksRetrieved,
-    ScoringRetrieved
-  };
-
   unsigned short fNThread{0};       ///< Number of G4 workers
   unsigned int fTrackCapacity{0};   ///< Number of track slots to allocate on device
   unsigned int fScoringCapacity{0}; ///< Number of hit slots to allocate on device
   int fDebugLevel{1};               ///< Debug level
-  std::vector<AdePTGeant4Integration> fG4Integrations;
-  std::unique_ptr<GPUstate> fGPUstate;               ///< CUDA state placeholder
-  std::vector<PerEventScoring> fScoring;             ///< User scoring objects per G4 worker
+  int fCUDAStackLimit{0};                              ///< CUDA device stack limit
+  std::vector<IntegrationLayer> fIntegrationLayerObjects;
+  // We can't use a unique_ptr because we can't have the definition of GPUState
+  // in a file compiled by gcc
+  GPUstate *fGPUstate;               ///< CUDA state placeholder
+  // We need to use pointers to the scoring objects because we can't have the complete definition 
+  // in a file compiled by gcc
+  std::vector<AdePTScoring*> fScoring;               ///< User scoring objects per G4 worker
   std::unique_ptr<TrackBuffer> fBuffer{nullptr};     ///< Buffers for transferring tracks between host and device
   std::unique_ptr<G4HepEmState> fg4hepem_state;      ///< The HepEm state singleton
   std::thread fGPUWorker;                            ///< Thread to manage GPU
@@ -74,29 +70,29 @@ private:
   bool fTrackInAllRegions = false;
   std::vector<std::string> const *fGPURegionNames;
 
-  void FullInit();
+  void Initialize();
   void InitBVH();
+  bool InitializeField(double bz);
   bool InitializeGeometry(const vecgeom::cxx::VPlacedVolume *world);
-  bool InitializePhysics(double bz);
-  void InitializeGPU();
+  bool InitializePhysics();
+  // void InitializeGPU();
   void FreeGPU();
-  /// @brief Asynchronous loop for transporting particles on GPU.
-  void TransportLoop();
-  void HitProcessingLoop(HitProcessingContext *const);
-  void ReturnTracksToG4();
-  void AdvanceEventStates(EventState oldState, EventState newState);
-  std::shared_ptr<const std::vector<GPUHit>> GetGPUHits(unsigned int threadId) const;
+  // /// @brief Asynchronous loop for transporting particles on GPU.
+  // void TransportLoop();
+  // void HitProcessingLoop(HitProcessingContext *const);
+  // void ReturnTracksToG4();
+  // void AdvanceEventStates(EventState oldState, EventState newState);
+  // std::shared_ptr<const std::vector<GPUHit>> GetGPUHits(unsigned int threadId) const;
 
 public:
-  AdeptIntegration(unsigned short nThread, unsigned int trackCapacity, unsigned int hitBufferCapacity, int debugLevel,
-                   std::vector<std::string> const *GPURegionNames, bool trackInAllRegions, int cudaStackSize);
-  AdeptIntegration(const AdeptIntegration &other) = delete;
-  ~AdeptIntegration();
+  AsyncAdePTTransport(AdePTConfiguration &configuration);
+  AsyncAdePTTransport(const AsyncAdePTTransport &other) = delete;
+  ~AsyncAdePTTransport();
 
   /// @brief Adds a track to the buffer
   void AddTrack(int pdg, int parentID, double energy, double x, double y, double z, double dirx, double diry,
                 double dirz, double globalTime, double localTime, double properTime, int threadId, unsigned int eventId,
-                unsigned int trackIndex) override;
+                unsigned int trackIndex, vecgeom::NavigationState &&state) override;
   /// @brief Set track capacity on GPU
   void SetTrackCapacity(size_t capacity) override { fTrackCapacity = capacity; }
   /// @brief Set Hit buffer capacity on GPU and Host
@@ -114,7 +110,13 @@ public:
   std::vector<std::string> const *GetGPURegionNames() override { return fGPURegionNames; }
   /// No effect
   void Initialize(bool) override {}
+  /// @brief Initializes the ApplyCut flag. Can only be called after G4 Physics is build
+  bool InitializeApplyCuts(bool applycuts);
   /// @brief Finish GPU transport, bring hits and tracks to host
+  /// @details The shower call exists to maintain the same interface as the 
+  /// synchronous AdePT mode, since in this case the transport loop is always 
+  /// running. The only call to Shower() from G4 is done when the tracking 
+  /// manager needs to flush an event.
   void Shower(int event, int threadId) override { Flush(threadId, event); }
   /// Block until transport of the given event is done.
   void Flush(int threadId, int eventId);
@@ -122,5 +124,7 @@ public:
 };
 
 } // namespace AsyncAdePT
+
+#include "AsyncAdePTTransport.icc"
 
 #endif
