+++ include/AdePT/core/AsyncTrack.cuh
@@ -1,22 +1,27 @@
 // SPDX-FileCopyrightText: 2022 CERN
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef ADEPT_TRACK_CUH
-#define ADEPT_TRACK_CUH
+#ifndef ASYNC_ADEPT_TRACK_CUH
+#define ASYNC_ADEPT_TRACK_CUH
 
 #include <AdePT/base/MParray.h>
 #include <AdePT/copcore/SystemOfUnits.h>
 #include <AdePT/copcore/Ranluxpp.h>
 
 #include <VecGeom/base/Vector3D.h>
-#include <VecGeom/navigation/NavStateIndex.h>
+#include <VecGeom/navigation/NavigationState.h>
+
+
+// TODO: This needs to be unified with the other Track struct, however due to the slot manager 
+// approach, this can't be done before introducing the SlotManager for all kernels
+
 
 // A data structure to represent a particle track. The particle type is implicit
 // by the queue and not stored in memory.
 struct Track {
   using Precision = vecgeom::Precision;
   RanluxppDouble rngState;
-  double energy            = 0;
+  double eKin            = 0;
   float numIALeft[3]       = {-1., -1., -1.};
   float initialRange       = -1.f; // Only for e-?
   float dynamicRangeFactor = -1.f; // Only for e-?
@@ -28,7 +33,7 @@ struct Track {
 
   vecgeom::Vector3D<Precision> pos;
   vecgeom::Vector3D<Precision> dir;
-  vecgeom::NavStateIndex navState;
+  vecgeom::NavigationState navState;
   unsigned int eventId{0};
   int parentId{-1};
   short threadId{-1};
@@ -42,7 +47,7 @@ struct Track {
   __device__ Track(uint64_t rngSeed, double eKin, double globalTime, float localTime, float properTime,
                    double const position[3], double const direction[3], unsigned int eventId, int parentId,
                    short threadId)
-      : energy{eKin}, globalTime{globalTime}, localTime{localTime}, properTime{properTime}, eventId{eventId},
+      : eKin{eKin}, globalTime{globalTime}, localTime{localTime}, properTime{properTime}, eventId{eventId},
         parentId{parentId}, threadId{threadId}
   {
     rngState.SetSeed(rngSeed);
@@ -53,10 +58,10 @@ struct Track {
 
   /// Construct a secondary from a parent track.
   /// NB: The caller is responsible to branch a new RNG state.
-  __device__ Track(RanluxppDouble const &rngState, double energy, const vecgeom::Vector3D<Precision> &parentPos,
-                   const vecgeom::Vector3D<Precision> &newDirection, const vecgeom::NavStateIndex &newNavState,
+  __device__ Track(RanluxppDouble const &rngState, double eKin, const vecgeom::Vector3D<Precision> &parentPos,
+                   const vecgeom::Vector3D<Precision> &newDirection, const vecgeom::NavigationState &newNavState,
                    const Track &parentTrack)
-      : rngState{rngState}, energy{energy}, globalTime{parentTrack.globalTime}, pos{parentPos}, dir{newDirection},
+      : rngState{rngState}, eKin{eKin}, globalTime{parentTrack.globalTime}, pos{parentPos}, dir{newDirection},
         navState{newNavState}, eventId{parentTrack.eventId}, parentId{parentTrack.parentId},
         threadId{parentTrack.threadId}
   {
