+++ include/AdePT/navigation/AdePTNavigator.h
@@ -9,13 +9,13 @@
 #ifndef ADEPT_NAVIGATOR_H_
 #define ADEPT_NAVIGATOR_H_
 
-#include <AdePT/navigation/LoopNavigator.h>
+// #include <AdePT/navigation/LoopNavigator.h>
 #include <AdePT/navigation/BVHNavigator.h>
 #ifdef ADEPT_USE_SURF
 #include <AdePT/navigation/SurfNavigator.h>
 #endif
 
-inline namespace COPCORE_IMPL {
+// inline namespace COPCORE_IMPL {
 #ifdef ADEPT_USE_SURF
 #ifdef ADEPT_USE_SURF_SINGLE
 using AdePTNavigator = SurfNavigator<float>;
@@ -25,5 +25,5 @@ using AdePTNavigator = SurfNavigator<double>;
 #else
 using AdePTNavigator = BVHNavigator;
 #endif
-} // End namespace COPCORE_IMPL
+// } // End namespace COPCORE_IMPL
 #endif // ADEPT_NAVIGATOR_H_
